Move hooks directory
