+++ components/ListOption.tsx
@@ -5,7 +5,7 @@ import classNames from 'classnames/bind';
 import { ListOption as ListOptionType } from '@/types/common';
 import { isOneOf } from '@/types/guard';
 import { SettingsPerPage } from '@/types/store/todoSlice';
-import { useAppDispatch, useAppSelector } from '@/hooks/index';
+import { useAppDispatch, useAppSelector } from '@/lib/hooks/index';
 import {
   CHANGE_THEME,
   TOGGLE_COMPLETED_ITEMS,
