+++ components/OrderingIndicator.tsx
@@ -3,7 +3,7 @@ import invariant from 'tiny-invariant';
 import classNames from 'classnames/bind';
 import { OrderingCriterion } from '@/types/common';
 import { isOneOf } from '@/types/guard';
-import { useAppDispatch, useAppSelector } from '@/hooks/index';
+import { useAppDispatch, useAppSelector } from '@/lib/hooks/index';
 import {
   IMPORTANCE,
   DEADLINE,
