+++ pages/tasks/planned.tsx
@@ -2,7 +2,7 @@ import Head from 'next/head';
 import classNames from 'classnames/bind';
 import dayjs from '@/lib/plugins/dayjs';
 import styles from './inbox.module.scss'; // shared
-import { useAppSelector } from '@/hooks/index';
+import { useAppSelector } from '@/lib/hooks/index';
 import PageToolbar from '@/components/PageToolbar';
 import TaskInput from '@/components/TaskInput';
 import TaskList from '@/components/TaskList';
